Massive Changes to separte out formbuilder.
Changed form builder so that it does not hold on to a specific class or form.
It is not more of a trnaslator for Forms.
Added Form which is for the forms.
Added unsupported Form Elements so now form can supprt elemenst that it does
not have and it will display its tostring which is
unediatable also the object is able to be gotten back.
Added Test classes for formbuilder, Formelements: int, double, string
Added annotation for Display name on the form.